requirements.txt
# 基本依赖
python-dotenv>=0.19.0
aiohttp>=3.8.1
requests==2.31.0
numpy>=1.21.0
scikit-learn>=1.0.0
beautifulsoup4>=4.10.0
webdriver-manager>=3.8.0
selenium>=4.1.0

# 不再需要 sentence-transformers
# sentence-transformers>=2.2.0

# 服务依赖
flask>=2.0.0
celery>=5.2.0